+++ packages/tswebextension/CHANGELOG.md
@@ -10,16 +10,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `filterId` query param to `documentBlockingPageUrl` for MV2.
+- Support for Chrome's User Scripts API that allows more reliable script
+  injection in MV3 if developer mode is enabled.
+- Static getter `isUserScriptsApiSupported` to check if the current browser
+  supports User Scripts API.
 
 ### Fixed
 
 - Invalid HTML rule selectors are breaking site loading
   [AdguardBrowserExtension#2646], [AdguardBrowserExtension#2826].
+- Scriptlet rules are not displayed in the filtering log [AdguardBrowserExtension#3164].
 - ya.ru - `Hide Referrer from third parties` breaks the site [AdguardBrowserExtension#2839].
 
 [AdguardBrowserExtension#2646]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2646
 [AdguardBrowserExtension#2826]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2826
 [AdguardBrowserExtension#2839]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2839
+[AdguardBrowserExtension#3164]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3164
 
 ## [3.0.1] - 2025-03-06